- Refactored code for improved readability and organization
