--- app/src/main/java/dev/mkao/weaver/domain/model/UppsertCase.kt
@@ -1,10 +1,0 @@
-package dev.mkao.weaver.domain.model
-
-import dev.mkao.weaver.data.remote.NewsDao
-
-class UppsertCase (private val newsDao: NewsDao){
-
-    suspend operator fun invoke(article: Article){
-        newsDao.upsert(article = article)
-    }
-}
