+++ app/src/main/java/dev/mkao/weaver/presentation/common/BottomNavigationBar.kt
@@ -1,9 +1,7 @@
-package dev.mkao.weaver.presentation.navgraph
+package dev.mkao.weaver.presentation.common
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.CheckCircle
-import androidx.compose.material.icons.filled.FavoriteBorder
-import androidx.compose.material.icons.filled.Home
 import androidx.compose.material.icons.filled.Settings
 import androidx.compose.material3.Icon
 import androidx.compose.material3.NavigationBar
