+++ app/src/main/java/dev/mkao/weaver/presentation/navigation/Screen.kt
@@ -1,12 +1,10 @@
-package dev.mkao.weaver.presentation.Bookmarks
+package dev.mkao.weaver.presentation.navigation
 
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.List
 import androidx.compose.material.icons.filled.CheckCircle
 import androidx.compose.material.icons.filled.FavoriteBorder
 import androidx.compose.material.icons.filled.Home
 import androidx.compose.material.icons.filled.Info
-import androidx.compose.material.icons.filled.List
 import androidx.compose.material.icons.filled.Settings
 import androidx.compose.ui.graphics.vector.ImageVector
